+++ src/main/java/com/sports/server/auth/utils/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.sports.server.auth.jwt;
+package com.sports.server.auth.utils;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
@@ -17,7 +17,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class JwtProvider {
+public class JwtUtil {
 
     @Value("${jwt.secret-key}")
     private String secretKey;
